+++ tests/src/lib.rs
@@ -467,7 +467,7 @@ fn run_tests(cfg: TestConfig) -> Result<bool, io::Error> {
                 tuple::raw_bytes,
             ]);
             tests.append(&mut tests![
-                [should_panic_if: !tarantool::ffi::has_fully_temporary_spaces()]
+                [should_panic_if: unsafe { !tarantool::ffi::has_fully_temporary_spaces() }]
                 r#box::fully_temporary_space,
             ]);
             tests.append(&mut tests![